backend/requirements.txt
# Web Framework & Server
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# Machine Learning & AI
torch>=2.9.1
transformers>=4.57.1
openai>=1.0.0

# Image Processing
pillow>=11.1.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
